requirements.txt
astroid==1.5.3
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
colorama==0.3.9
Flask==0.12.2
gevent==1.2.2
greenlet==0.4.12
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
locustio==0.8.1
MarkupSafe==1.0
mccabe==0.6.1
msgpack-python==0.4.8
pylint==1.7.2
pyzmq==16.0.2
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
wrapt==1.10.11
